+++ GUI/package.json
@@ -91,8 +91,8 @@
               "config": "./webpack.renderer.config.js",
               "entryPoints": [
                 {
-                  "html": "./src/index.html",
-                  "js": "./src/renderer.ts",
+                  "html": "./src/renderer/index.html",
+                  "js": "./src/renderer/renderer.ts",
                   "name": "main_window"
                 }
               ]
