+++ GUI/src/renderer/index.html
@@ -1,0 +1,140 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+  </head>
+  <body>
+    <header class="mdc-top-app-bar">
+      <div class="mdc-top-app-bar__row">
+        <section class="mdc-top-app-bar__section mdc-top-app-bar__section--align-start">
+          <span class="mdc-top-app-bar__title">Scheduler</span>
+          <span id="runningStatus" class="running mdc-typography mdc-top-app-bar__title"></span>
+        </section>
+        <section class="mdc-top-app-bar__section mdc-top-app-bar__section--align-end" role="toolbar">
+          <button
+            id="minimize"
+            class="material-icons mdc-top-app-bar__action-item mdc-icon-button"
+            aria-label="Favorite"
+          >
+            minimize
+          </button>
+          <button id="maximize" class="material-icons mdc-top-app-bar__action-item mdc-icon-button" aria-label="Search">
+            maximize
+          </button>
+          <button id="close" class="material-icons mdc-top-app-bar__action-item mdc-icon-button" aria-label="Options">
+            close
+          </button>
+        </section>
+      </div>
+      <div
+        id="loader"
+        role="progressbar"
+        class="mdc-linear-progress mdc-linear-progress--indeterminate mdc-linear-progress--closed"
+        aria-label="Example Progress Bar"
+        aria-valuemin="0"
+        aria-valuemax="1"
+        aria-valuenow="0"
+      >
+        <div class="mdc-linear-progress__buffer">
+          <div class="mdc-linear-progress__buffer-bar"></div>
+          <div class="mdc-linear-progress__buffer-dots"></div>
+        </div>
+        <div class="mdc-linear-progress__bar mdc-linear-progress__primary-bar">
+          <span class="mdc-linear-progress__bar-inner"></span>
+        </div>
+        <div class="mdc-linear-progress__bar mdc-linear-progress__secondary-bar">
+          <span class="mdc-linear-progress__bar-inner"></span>
+        </div>
+      </div>
+    </header>
+
+    <main class="mdc-top-app-bar--fixed-adjust">
+      <aside id="schedule"></aside>
+      <div>
+        <section id="clock"></section>
+        <section id="newTask">
+          <form id="taskForm">
+            <div>
+              <div class="row">
+                <!-- Task -->
+                <label id="task" class="mdc-text-field mdc-text-field--outlined">
+                  <span class="mdc-notched-outline">
+                    <span class="mdc-notched-outline__leading"></span>
+                    <span class="mdc-notched-outline__notch">
+                      <span class="mdc-floating-label" id="my-label-id">Task</span>
+                    </span>
+                    <span class="mdc-notched-outline__trailing"></span>
+                  </span>
+                  <input
+                    id="command"
+                    type="text"
+                    class="mdc-text-field__input"
+                    aria-labelledby="my-label-id"
+                    name="command"
+                  />
+                </label>
+                <button class="mdc-button mdc-button--raised" id="browse">
+                  <span class="mdc-button__label">Browse...</span>
+                </button>
+              </div>
+
+              <div class="row">
+                <!-- Name -->
+                <label id="name" class="mdc-text-field mdc-text-field--outlined">
+                  <span class="mdc-notched-outline">
+                    <span class="mdc-notched-outline__leading"></span>
+                    <span class="mdc-notched-outline__notch">
+                      <span class="mdc-floating-label" id="my-label-id">Name</span>
+                    </span>
+                    <span class="mdc-notched-outline__trailing"></span>
+                  </span>
+                  <input type="text" class="mdc-text-field__input" aria-labelledby="my-label-id" name="name" />
+                </label>
+
+                <!-- interval -->
+                <label id="interval" class="mdc-text-field mdc-text-field--outlined">
+                  <span class="mdc-notched-outline">
+                    <span class="mdc-notched-outline__leading"></span>
+                    <span class="mdc-notched-outline__notch">
+                      <span class="mdc-floating-label" id="my-label-id">Interval</span>
+                    </span>
+                    <span class="mdc-notched-outline__trailing"></span>
+                  </span>
+                  <input type="text" class="mdc-text-field__input" aria-labelledby="my-label-id" name="interval" />
+                </label>
+              </div>
+
+              <div class="row">
+                <!-- Description -->
+                <label
+                  id="description"
+                  class="mdc-text-field mdc-text-field--outlined mdc-text-field--textarea mdc-text-field--no-label"
+                >
+                  <span class="mdc-notched-outline">
+                    <span class="mdc-notched-outline__leading"></span>
+                    <span class="mdc-notched-outline__notch">
+                      <span class="mdc-floating-label" id="my-label-id">Description</span>
+                    </span>
+                    <span class="mdc-notched-outline__trailing"></span>
+                  </span>
+                  <span class="container">
+                    <textarea class="mdc-text-field__input" rows="4" aria-label="Label" name="description"></textarea>
+                  </span>
+                </label>
+              </div>
+            </div>
+            <div class="mdc-touch-target-wrapper" id="addTask">
+              <button class="mdc-fab mdc-fab--extended">
+                <div class="mdc-fab__ripple"></div>
+                <span class="material-icons mdc-fab__icon">add</span>
+                <span class="mdc-fab__label">Create</span>
+              </button>
+            </div>
+          </form>
+        </section>
+      </div>
+    </main>
+  </body>
+</html>
